Creating a Xamarin iOS binding for ZipArchive
The sample does run well as the folder it attempts to zip does not exist.

The ZipFolder method should be checked as I did a port and did not check it closely.

The was written to match the current component on the store.  The event and delegates in the binding are not 100% correct, but the match what is there now.